requirements.txt
# Django and core
Django==5.2.4
djangorestframework
djangorestframework-simplejwt
djangorestframework-simplejwt[crypto]
django-rest-framework-simplejwt-blacklist
django-allauth
dj-rest-auth

python-decouple==3.8
python-slugify==8.0.4
Pillow==11.3.0
PyJWT==2.10.1
Requests==2.32.4
weasyprint
django-widget-tweaks
django-debug-toolbar
cryptography
psycopg2

# Redis and Celery
celery==5.4.0
redis==5.0.4
django-redis


factory_boy==3.3.3
Faker==37.4.2

# Auth and features
django-allauth==65.10.0
features==0.5.12
gunicorn
